+++ src/mjolnir/graphtilebuilder.cc
@@ -1,3 +1,4 @@
+#include "baldr/predictedspeeds.h"
 #include "mjolnir/graphtilebuilder.h"
 #include "baldr/directededge.h"
 #include "baldr/edgeinfo.h"
@@ -1255,14 +1256,26 @@ void GraphTileBuilder::AddPredictedSpeed(const uint32_t idx,
   if (speed_profile_offset_builder_.size() == 0) {
     speed_profile_offset_builder_.resize(header_->directededgecount());
     speed_profile_builder_.reserve(predicted_count_hint * kCoefficientCount);
+    speed_profile_map_.reserve(predicted_count_hint);
   }
 
+  // If the speed profile exists, reuse the offset for this directed edge.
+  auto it = speed_profile_map_.find(std::span{coefficients});
+  if (it != speed_profile_map_.end()) {
+    speed_profile_offset_builder_[idx] = it->second;
+    return;
+  }
+
+  auto new_speed_profile_offset = static_cast<uint32_t>(speed_profile_builder_.size());
+
   // Set the offset to the predicted speed profile for this directed edge
-  speed_profile_offset_builder_[idx] = speed_profile_builder_.size();
+  speed_profile_offset_builder_[idx] = new_speed_profile_offset;
 
   // Append the profile
   speed_profile_builder_.insert(speed_profile_builder_.end(), coefficients.begin(),
                                 coefficients.end());
+
+  speed_profile_map_.emplace(std::span{speed_profile_builder_.data() + new_speed_profile_offset, kCoefficientCount}, new_speed_profile_offset);
 }
 
 // Updates a tile with predictive speed data. Also updates directed edges with